+++ src/mock/film.js
@@ -1,5 +1,5 @@
 import { getRandomArrayElement } from '../utils.js';
-import { POSTERS, GENRES, EMOTIONS } from '../const.js';
+import { POSTERS, GENRES, EMOTIONS } from '../consts.js';
 
 const mockComments = [
   {
